requirements.txt
# Web Framework
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6

# Machine Learning
tensorflow>=2.15.0
scikit-learn>=1.3.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
pillow>=10.0.0

# Database
sqlalchemy>=2.0.0

# Visualization (for Grad-CAM)
matplotlib>=3.7.0

# Kaggle API (for dataset download)
kaggle>=1.5.0

# Development
python-dotenv>=1.0.0
